pyproject.toml
[project]
name = "chat-with-my-database"
version = "0.1.1"
description = "Use natural language to query SQLite database. Powered by OpenAI's GPT models and Pydantic AI."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "pandas>=2.2.3",
    "pydantic-ai-slim[openai]>=0.8.0",
    "streamlit>=1.44.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pylint>=3.3.6",
    "pyright>=1.1.398",
    "pytest>=8.3.5",
    "ruff>=0.11.2",
]
